+++ dart/datatools/lib/src/file/fetch/file_content.dart
@@ -89,7 +89,4 @@ class FileContent extends Content {
       throw ClientException.decodingJsonFailed(e);
     }
   }
-
-  @override
-  Future<Stream<List<int>>> get stream async => byteStream();
 }