+++ dart/datatools/lib/src/http/fetch/http_content.dart
@@ -114,7 +114,4 @@ class HttpContent extends Content {
       throw ClientException.decodingJsonFailed(e);
     }
   }
-
-  @override
-  Future<Stream<List<int>>> get stream async => byteStream();
 }